requirements.txt
supabase
python-dotenv
rich